[IOTDB-5410] Refine IT: Refine Session and SessionPool IT
